requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
numpy
scikit-learn
openai
